server/requirements.txt
langchain==0.0.340
openai==1.3.4
PyYAML==6.0.1
tiktoken==0.7.0
Flask>=3.0.3

# Voiceover Requirements
elevenlabs==1.9.0
python-dotenv==1.0.1
# Voiceover dev purposes
pydub==0.25.1
simpleaudio==1.0.4
